added CirculationController, moved circ to "models" directory
